backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary>=2.9.10

# CORS for frontend communication
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Image processing (for future product recognition)
Pillow>=10.4.0

# Environment variables
python-decouple==3.8

# Development tools
django-extensions==3.2.3

# API documentation
drf-spectacular==0.26.5

# Timezone support
pytz==2023.3

# Filtering for API
django-filter==23.2
